requirements.txt
# Core Dependencies
google-genai==1.0.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
pydantic>=2.0.0
numpy<2.0.0

# Search & Memory
chromadb==0.4.24
langchain-text-splitters>=0.0.1
supabase>=2.0.0
vecs>=0.4.0

# Utilities
ruamel.yaml==0.18.0
python-dotenv==1.0.0
colorama==0.4.6
tenacity==8.5.0
watchdog>=4.0.0
markdown>=3.0.0

# Development & Testing
ruff==0.6.0
pytest==8.3.0
pytest-asyncio==0.23.0
